feat: Add TaskSection component
This commit adds the TaskSection component, which is responsible for rendering a section of tasks. It includes the ability to expand and collapse the section, as well as display a list of tasks. The component uses React Native's Animated API to animate the height of the section when it is expanded or collapsed.