feat: Add update and delete functionality to TaskContext
This commit adds the update and delete functionality to the TaskContext. It includes the implementation of the update and delete methods in the TaskContextProvider component, as well as the necessary changes in the TaskRepositoryHttp class. These additions allow users to update and delete tasks, enhancing the overall functionality of the application.