chore: Remove react-native-dotenv package and update dependencies
This commit removes the react-native-dotenv package and updates the dependencies in the package.json and yarn.lock files. The removal of react-native-dotenv is done to keep the project up to date with the latest stable versions of the dependencies and remove any unused packages. This chore improves the project's maintainability and ensures a clean dependency tree.