feat: Add task_id property to TaskCardProps interface
This commit adds a new property called `task_id` to the `TaskCardProps` interface in the `TaskCard` component. The `task_id` property is of type `string` and allows for passing the ID of a task to the `TaskCard` component. This change enhances the functionality of the `TaskCard` component by providing a way to associate tasks with their respective IDs.